Now I have a port on east side. Go team
First commit of the morning. Buzzing like crazy. We reflected on
yesterday that maybe we were a bit too 'sustainable' with the breaks.
Also both of us work at different time zones (no really).

It's now 7:26am and I am caffienated and raring to go.

Still feel I am not using clojure as well as I could but I am sticking
with it because the language isn't causing me problems (the biggest
problems I have are spatial coordintes and understanding craftyjs).

I think that the entity-component stuff is actually a good fit -
treating stuff as data. It's just that crafty is too javascripty. The
"this-as" stuff everywhere I believe is a bit of a smell.

Anyway, I can see that there is a way to make stuff nicer. I am probably
not goingto do it unless it saves time (like doing so for coordinates
for ports).