Ports on the NSEW sides
Well this effectively makes https://trello.com/c/vJMK7Gd2 done.

Kinda glad my effort didn't require a lot of graphics. But a bit sad it
looks so crap though.

Next is to start emitting packets. It would be nice to have a state for
these though to indicate when they are active or not (initial level
might just have one).