Create packets from a port on activation
When a port activates, it emits 10 packets on a delay timer then stops.
Each packet create raises an event.

The delay is cancelled when the port is destroyed (ie the scene is
destroyed)

Learnt that the only things destroyed with a scene are the 2d objects.
So made packet a 2d object. Not sure this is best way but oh wells.