Basic box bounds done in horrible way
Now when packets try to go out of bounds, we have something to tell them
not to. A series of polygons around the edge. Kinda gross. But it seems
to work.

If wanted to do the same with the ports too need to add on a little
grace period for them to become free of the box - or change the point
where they come out.