Now the packets stay in the box
It was hard earned but... I made it! Quite happy that I got there, not
so much about all the faffing.

Feel I shouldn't have to be doing so much with angles. A look for
libraries and stuff didn't yield great results.

Now I done this... I think it's time to start making this a game!