Now uses a queue to store packets and checks from router
Learnt that state on the game scene isn't awesome as you have to unbind
events etc. Seems like at least for entities in the scene they are
nicely disposed of. So now router does that checking for us (makes it
equally easy to have more than one router I guess !)

Packets are in a queue so I can get a better solution for painting the
'first' one