Changed the make-components(-awesome) function
This change removes the need for all the init functions for components
to specify their 'requires' - that is the components that are bolted in
to provide this component's behaviour.

At the same time, made the init fn take in a 'me' to remove but one more
this-as me call from a bunch of places.

It's a little bit dumb still, but it's a good start and giving me a good
idea of where to progress with it.

Current thinking would be to remove all the extra functions that are
bolted in on last arg to make-components-awesome, make them pure and
invoke them from clojurescript directly.

I think eventually i will have to think about event handling and
dispatch but it would be nice to do that cleanly. Still need to grok
both the current implementation of the entity-component pattern and
maybe more standard interpretations of it (entity-component-system ?)